requirements.txt
# Django
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.0
django-filter==23.3

# Database - Using newer version that works better on Windows
psycopg2-binary==2.9.10

# Authentication & Security
djangorestframework-simplejwt==5.3.0
django-allauth==0.57.0
python-decouple==3.8

# File handling - Updated Pillow version
Pillow==10.3.0
python-magic-bin==0.4.14

# API Documentation
drf-yasg==1.21.7

# Async tasks
celery==5.3.4
redis==5.0.1
django-celery-beat==2.5.0
django-celery-results==2.5.1

# Storage (for production)
django-storages==1.14.2
boto3==1.34.0

# PDF Generation
reportlab==4.1.0
# weasyprint==60.1  # Commented out - can cause issues on Windows, uncomment if needed

# Excel/CSV
openpyxl==3.1.2
pandas==2.1.4

# Email
# django-ses==3.5.2  # Commented out - optional, uncomment if using AWS SES

# Monitoring & Logging
sentry-sdk==1.40.0
django-debug-toolbar==4.2.0

# Testing
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
factory-boy==3.3.0
faker==22.0.0

# Code quality
black==24.1.0
flake8==7.0.0
isort==5.13.0
# pylint==3.0.3  # Commented out - optional

# Production server
gunicorn==21.2.0
whitenoise==6.6.0

# Utilities
python-dateutil==2.8.2
pytz==2024.1
